@@ -120,9 +120,9 @@ describe 'action_column WiceGrid', type: :request, js: true do
   end
 
   it 'should filter by Added inside a form' do
-    set_datepicker(self, 'g_f_created_at_fr_date_placeholder', 2011, 5, 1)
+    set_datepicker(self, 'g_f_created_at_fr_date_placeholder', 2021, 5, 1)
 
-    set_datepicker(self, 'g_f_created_at_to_date_placeholder', 2011, 9, 1)
+    set_datepicker(self, 'g_f_created_at_to_date_placeholder', 2021, 9, 1)
 
     first(:css, 'button.btn', text: 'Process tasks').click
 
@@ -131,7 +131,7 @@ describe 'action_column WiceGrid', type: :request, js: true do
     end
 
     within 'div.wice-grid-container table.wice-grid tbody tr:first-child td.active-filter' do
-      expect(page).to have_content('2011-09-13 22:11:12')
+      expect(page).to have_content('2021-09-13 22:11:12')
     end
 
     within 'div.wice-grid-container table.wice-grid thead' do
@@ -139,7 +139,7 @@ describe 'action_column WiceGrid', type: :request, js: true do
     end
 
     within 'div.wice-grid-container table.wice-grid tbody tr:first-child td.active-filter' do
-      expect(page).to have_content('2011-09-29 22:11:12')
+      expect(page).to have_content('2021-09-29 22:11:12')
     end
 
     within '.pagination_status' do
@@ -151,15 +151,15 @@ describe 'action_column WiceGrid', type: :request, js: true do
     end
 
     within 'div.wice-grid-container table.wice-grid tbody tr:first-child td.active-filter' do
-      # page.should have_content('2011-08-14 22:11:12')
-      expect(page).to have_content('2011-09-22 22:11:12')
+      # page.should have_content('2021-08-14 22:11:12')
+      expect(page).to have_content('2021-09-22 22:11:12')
     end
   end
 
   it 'should filter by Due Date' do
-    set_datepicker(self, 'g_f_due_date_fr_date_placeholder', 2012, 0, 1)
+    set_datepicker(self, 'g_f_due_date_fr_date_placeholder', 2022, 0, 1)
 
-    set_datepicker(self, 'g_f_due_date_to_date_placeholder', 2013, 0, 1)
+    set_datepicker(self, 'g_f_due_date_to_date_placeholder', 2023, 0, 1)
 
     first(:css, 'button.btn', text: 'Process tasks').click
 
@@ -168,7 +168,7 @@ describe 'action_column WiceGrid', type: :request, js: true do
     end
 
     within 'div.wice-grid-container table.wice-grid tbody tr:first-child td.active-filter' do
-      expect(page).to have_content('2012-07-29')
+      expect(page).to have_content('2022-07-29')
     end
 
     within 'div.wice-grid-container table.wice-grid thead' do
@@ -176,7 +176,7 @@ describe 'action_column WiceGrid', type: :request, js: true do
     end
 
     within 'div.wice-grid-container table.wice-grid tbody tr:first-child td.active-filter' do
-      expect(page).to have_content('2012-10-02')
+      expect(page).to have_content('2022-10-02')
     end
 
     within '.pagination_status' do
@@ -192,12 +192,12 @@ describe 'action_column WiceGrid', type: :request, js: true do
     end
 
     within 'div.wice-grid-container table.wice-grid tbody tr:first-child td.active-filter' do
-      expect(page).to have_content('2012-07-02')
+      expect(page).to have_content('2022-07-02')
     end
 
-    set_datepicker(self, 'g_f_due_date_fr_date_placeholder', 2012, 6, 28)
+    set_datepicker(self, 'g_f_due_date_fr_date_placeholder', 2022, 6, 28)
 
-    set_datepicker(self, 'g_f_due_date_to_date_placeholder', 2012, 6, 31)
+    set_datepicker(self, 'g_f_due_date_to_date_placeholder', 2022, 6, 31)
 
     first(:css, 'button.btn', text: 'Process tasks').click
 
