@@ -23,14 +23,14 @@ describe 'auto reloads 2 WiceGrid', type: :feature, js: true do
   end
 
   it 'should filter by Added' do
-    set_datepicker(self, 'grid_f_created_at_fr_date_placeholder', 2011, 5, 1)
+    set_datepicker(self, 'grid_f_created_at_fr_date_placeholder', 2021, 5, 1)
     within 'div.wice-grid-container table.wice-grid tbody tr:first-child td.active-filter' do
-      expect(page).to have_content('2011-09-13 22:11:12')
+      expect(page).to have_content('2021-09-13 22:11:12')
     end
 
-    set_datepicker(self, 'grid_f_created_at_to_date_placeholder', 2011, 9, 1)
+    set_datepicker(self, 'grid_f_created_at_to_date_placeholder', 2021, 9, 1)
     within 'div.wice-grid-container table.wice-grid tbody tr:first-child td.active-filter' do
-      expect(page).to have_content('2011-09-13 22:11:12')
+      expect(page).to have_content('2021-09-13 22:11:12')
     end
 
     within '.pagination_status' do
@@ -42,7 +42,7 @@ describe 'auto reloads 2 WiceGrid', type: :feature, js: true do
     end
 
     within 'div.wice-grid-container table.wice-grid tbody tr:first-child td.active-filter' do
-      expect(page).to have_content('2011-09-29 22:11:12')
+      expect(page).to have_content('2021-09-29 22:11:12')
     end
 
     within '.pagination_status' do
@@ -54,7 +54,7 @@ describe 'auto reloads 2 WiceGrid', type: :feature, js: true do
     end
 
     within 'div.wice-grid-container table.wice-grid tbody tr:first-child td.active-filter' do
-      expect(page).to have_content('2011-09-22 22:11:12')
+      expect(page).to have_content('2021-09-22 22:11:12')
     end
 
     find(:css, '.wg-external-reset-button').click
@@ -239,14 +239,14 @@ describe 'auto reloads 2 WiceGrid', type: :feature, js: true do
   end
 
   it 'should filter by Due Date' do
-    set_datepicker(self, 'grid_f_due_date_fr_date_placeholder', 2012, 0, 1)
+    set_datepicker(self, 'grid_f_due_date_fr_date_placeholder', 2022, 0, 1)
     within 'div.wice-grid-container table.wice-grid tbody tr:first-child td.active-filter' do
-      expect(page).to have_content('2013-01-26')
+      expect(page).to have_content('2023-01-26')
     end
 
-    set_datepicker(self, 'grid_f_due_date_to_date_placeholder', 2013, 0, 1)
+    set_datepicker(self, 'grid_f_due_date_to_date_placeholder', 2023, 0, 1)
     within 'div.wice-grid-container table.wice-grid tbody tr:first-child td.active-filter' do
-      expect(page).to have_content('2012-07-29')
+      expect(page).to have_content('2022-07-29')
     end
 
     within '.pagination_status' do
@@ -258,7 +258,7 @@ describe 'auto reloads 2 WiceGrid', type: :feature, js: true do
     end
 
     within 'div.wice-grid-container table.wice-grid tbody tr:first-child td.active-filter' do
-      expect(page).to have_content('2012-10-02')
+      expect(page).to have_content('2022-10-02')
     end
 
     within '.pagination_status' do
@@ -270,17 +270,17 @@ describe 'auto reloads 2 WiceGrid', type: :feature, js: true do
     end
 
     within 'div.wice-grid-container table.wice-grid tbody tr:first-child td.active-filter' do
-      expect(page).to have_content('2012-07-02')
+      expect(page).to have_content('2022-07-02')
     end
 
-    set_datepicker(self, 'grid_f_due_date_fr_date_placeholder', 2012, 6, 28)
+    set_datepicker(self, 'grid_f_due_date_fr_date_placeholder', 2022, 6, 28)
     within 'div.wice-grid-container table.wice-grid tbody tr:first-child td.active-filter' do
-      expect(page).to have_content('2012-10-02')
+      expect(page).to have_content('2022-10-02')
     end
 
-    set_datepicker(self, 'grid_f_due_date_to_date_placeholder', 2012, 6, 31)
+    set_datepicker(self, 'grid_f_due_date_to_date_placeholder', 2022, 6, 31)
     within 'div.wice-grid-container table.wice-grid tbody tr:first-child td.active-filter' do
-      expect(page).to have_content('2012-07-29')
+      expect(page).to have_content('2022-07-29')
     end
 
     within '.pagination_status' do
