@@ -72,14 +72,14 @@ describe 'auto reloads 3 WiceGrid', type: :request, js: true do
   # end
 
   it 'should filter by Due Date independantly' do
-    set_datepicker(self, 'grid_f_due_date_fr_date_placeholder', 2012, 0, 1)
+    set_datepicker(self, 'grid_f_due_date_fr_date_placeholder', 2022, 0, 1)
     within 'div#grid.wice-grid-container table.wice-grid tbody tr:first-child td.active-filter' do
-      expect(page).to have_content('2013-01-26')
+      expect(page).to have_content('2023-01-26')
     end
 
-    set_datepicker(self, 'grid_f_due_date_to_date_placeholder', 2013, 0, 1)
+    set_datepicker(self, 'grid_f_due_date_to_date_placeholder', 2023, 0, 1)
     within 'div#grid.wice-grid-container table.wice-grid tbody tr:first-child td.active-filter' do
-      expect(page).to have_content('2012-07-29')
+      expect(page).to have_content('2022-07-29')
     end
 
     within 'div#grid.wice-grid-container .pagination_status' do
@@ -90,14 +90,14 @@ describe 'auto reloads 3 WiceGrid', type: :request, js: true do
       expect(page).to have_content('1-20 / 50')
     end
 
-    set_datepicker(self, 'grid2_f_due_date_fr_date_placeholder', 2012, 0, 1)
+    set_datepicker(self, 'grid2_f_due_date_fr_date_placeholder', 2022, 0, 1)
     within 'div#grid2.wice-grid-container table.wice-grid tbody tr:first-child td.active-filter' do
-      expect(page).to have_content('2013-01-26')
+      expect(page).to have_content('2023-01-26')
     end
 
-    set_datepicker(self, 'grid2_f_due_date_to_date_placeholder', 2013, 0, 1)
+    set_datepicker(self, 'grid2_f_due_date_to_date_placeholder', 2023, 0, 1)
     within 'div#grid2.wice-grid-container table.wice-grid tbody tr:first-child td.active-filter' do
-      expect(page).to have_content('2013-01-26')
+      expect(page).to have_content('2023-01-26')
     end
 
     within 'div#grid.wice-grid-container .pagination_status' do
