@@ -23,21 +23,21 @@ describe 'dates WiceGrid', type: :request, js: true do
   include_examples 'Created At standard filtering'
 
   it 'should filter by multiple fields' do
-    select '2011', from: 'grid_f_updated_at_fr_year'
+    select '2021', from: 'grid_f_updated_at_fr_year'
     select 'January', from: 'grid_f_updated_at_fr_month'
     select '8', from: 'grid_f_updated_at_fr_day'
     select '00', from: 'grid_f_updated_at_fr_hour'
     select '00', from: 'grid_f_updated_at_fr_minute'
 
-    select '2011', from: 'grid_f_updated_at_to_year'
+    select '2021', from: 'grid_f_updated_at_to_year'
     select 'December', from: 'grid_f_updated_at_to_month'
     select '10', from: 'grid_f_updated_at_to_day'
     select '00', from: 'grid_f_updated_at_to_hour'
     select '00', from: 'grid_f_updated_at_to_minute'
 
-    set_datepicker(self, 'grid_f_created_at_fr_date_placeholder', 2011, 8, 1)
+    set_datepicker(self, 'grid_f_created_at_fr_date_placeholder', 2021, 8, 1)
 
-    set_datepicker(self, 'grid_f_created_at_to_date_placeholder', 2011, 8, 30)
+    set_datepicker(self, 'grid_f_created_at_to_date_placeholder', 2021, 8, 30)
 
     find(:css, '#grid_submit_grid_icon').click
 
