@@ -1,7 +1,6 @@
 require 'acceptance_helper'
 
 describe "upper pagination panel WiceGrid", :type => :request do
-
   before :each do
     visit '/upper_pagination_panel'
   end
@@ -12,12 +11,11 @@ describe "upper pagination panel WiceGrid", :type => :request do
   end
 
   it "upper pagination panel should be present" do
+    skip "This test should be run with BROWSER=y in environment" unless ENV['BROWSER']
     expect(page).to have_selector('table.wice-grid thead tr td .pagination li.active')
 
     within 'table.wice-grid thead tr td .pagination_status' do
       expect(page).to have_content('1-20 / 50')
     end
   end
-
-
 end
