@@ -20,7 +20,7 @@ Examples::TestApp.configure do
   # Only use best-standards-support built into browsers
   config.action_dispatch.best_standards_support = :builtin
 
-  config.action_dispatch.show_exceptions = true
+  config.action_dispatch.show_exceptions = :rescuable
 
   # Do not compress assets
   config.assets.compress = false
